Working on Starting enviorment
Today I worked on the starting environment, I got a background city mostly done in terms of layout, I still need to color it and make a mountain to put the tunnel into. As well as work on the terrain surrounding the city. The roads might need to be changed or custom made because the ones currently being used are a little silly in the bad way.